@@@ コンパイル用のmake.shファイルの作成 @@@

+++ 00.Document/01.CMake/make.sh
@@ -2,17 +2,25 @@
 
 CUR_DIR=`pwd`
 DIR=""
+CMAKE_PREFIX_PATH=""
+CLEANUP=1
 
 if [[ $# -lt 1 ]];then
-    echo "Command Format:`basename $0` [-d dir]"
+    echo "Command Format:`basename $0` -d dir [-m cmake prefix path] [-c cleanup]"
     exit 1
 fi
 
-while getopts d: opt; do  
+while getopts d:m:c: opt; do  
 case $opt in   
 	d)  
 		DIR=$OPTARG
-		;;  
+		;; 
+        m)
+                CMAKE_PREFIX_PATH=$OPTARG
+                ;;
+        c)
+                CLEANUP=$OPTARG
+                ;;
 	*)  
 		;;  
 esac  
@@ -25,15 +33,21 @@ fi
 
 cd $DIR
 
-cmake .
+if [ ${CMAKE_PREFIX_PATH} == "" ]; then
+  cmake .
+else
+  cmake . -DCMAKE_PREFIX_PATH=${CMAKE_PREFIX_PATH}
+fi
 
 #Complie
 make
 
 # Clean
-rm -f ./CMakeCache.txt
-rm -rf ./CMakeFiles/
-rm -f ./Makefile
-rm -f ./cmake_install.cmake
+if [ ${CLEANUP} == "1" ]; then
+  rm -f ./CMakeCache.txt
+  rm -rf ./CMakeFiles/
+  rm -f ./Makefile
+  rm -f ./cmake_install.cmake
+fi
 
 cd $CUR_DIR
